python-code/api/requirement.txt
pandas==2.3.0
python-dotenv==1.0.1
openai==1.50.2
pinecone==5.3.1
runpod==1.7.1